ApertureService/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
openai==1.12.0
pyyaml==6.0.1
python-dotenv==1.0.1
aperturedb==0.9.0
torch==2.1.2
clip==1.0
numpy>=1.21.0
